+++ sumologic/data_source_sumologic_role.go
@@ -7,7 +7,7 @@ import (
 	"log"
 	"net/url"
 
-	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
 func dataSourceSumologicRole() *schema.Resource {