+++ sumologic/data_sumologic_source_caller_identity.go
@@ -1,7 +1,7 @@
 package sumologic
 
 import (
-	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
 func dataSourceSumologicCallerIdentity() *schema.Resource {