+++ sumologic/resource_sumologic_cse_network_block.go
@@ -2,7 +2,7 @@ package sumologic
 
 import (
 	"errors"
-	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"log"
 )