+++ sumologic/resource_sumologic_extraction_rule.go
@@ -15,7 +15,7 @@ import (
 	"log"
 	"strings"
 
-	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
 func resourceSumologicFieldExtractionRule() *schema.Resource {